shes_backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-ratelimit==4.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
cryptography==42.0.8
Pillow==10.4.0
django-filter==24.3
dj-database-url==2.2.0
coverage==7.6.1
factory-boy==3.3.1
Faker==28.4.1
